+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Release Notes for 1.x
 
+
 ## [Upcoming Release]
 
 - This release introduces the new feature of automatic release creation.
@@ -11,6 +12,12 @@
 ## [Automated Release Notes]
 
 - Placeholder for automated release notes
+
+## 2.0.0 - November 15, 2024
+
+- ✨ Added new ``wpkirk_md` helper function to parse the backticks in the string.
+- ✨ Improved the all helpers functions.
+- Removed the useless functions.
 
 ## 1.0.0 - October 15, 2024